+++ api/pyproject.toml
@@ -26,6 +26,8 @@ pytest-asyncio = "^0.23.8"
 flake8 = "^7.1.1"
 black = "^24.8.0"
 isort = "^5.13.2"
+aiosqlite = "^0.20.0"
+httpx = "^0.27.2"
 
 [build-system]
 requires = ["poetry-core"]
