+++ api/tests/test_v2_user.py
@@ -1,31 +1,15 @@
+import starlette.status
 import pytest
-from httpx import AsyncClient
-from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
-from sqlalchemy.orm import sessionmaker
-
-from db import get_db, Base
-from main import app
-from v2.utils.config import get_async_db_url
-
-ASYNC_DB_URL = get_async_db_url()
 
 
-@pytest.fixture
-async def async_client():
-    async_engine = create_async_engine(ASYNC_DB_URL, echo=True)
-    async_session = sessionmaker(
-        autocommit=False, autoflush=False, bind=async_engine, class_=AsyncSession
+@pytest.mark.asyncio
+async def test_create_and_read(async_client):
+    response = await async_client.post(
+        "/v2/users/",
+        json={
+            "raspi_id": 1
+        }
     )
-
-    async with async_engine.begin() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
-        await conn.run_sync(Base.metadata.create_all)
-
-    async def get_test_db():
-        async with async_session() as session:
-            yield session
-
-    app.dependency_overrides[get_db] = get_test_db
-
-    async with AsyncClient(app=app, base_url="http://test") as client:
-        yield client
+    assert response.status_code == starlette.status.HTTP_200_OK
+    response_obj = response.json()
+    assert response_obj["raspi_id"] == 1
